// src/domain/models/serieItem.ts
/**
 * API de finanças M&J
 * API de controle financeiro do Júnior e da Vivi
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface SerieItem {
  mesReferencia: string;
  totalAtivos: number;
  totalPassivos: number;
  patrimonioLiquido: number;
}